+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - PIM-9755: Hide product model proposal changes that are already reviewed
 - PIM-9764: Fix open Asset in new tab in Asset Manager library
 - PIM-9767: Fix minimum & maximum user password validation
+- PIM-9768: Handle error when trying to delete a product model with at least 1 variant already published
 
 ## Improvements